% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_cleanup.R
\name{chebyshev_ineq_sample}
\alias{chebyshev_ineq_sample}
\title{Chebyshev Inequality with sample mean.
Utilizes Chebyshev inequality with sample mean.}
\usage{
chebyshev_ineq_sample(n, k)
}
\arguments{
\item{n}{Sample size}

\item{k}{Number of standard deviations.}
}
\description{
Reports the maximum percentage beyond k * sqrt((n + 1)/n)
standard deviations from mean
}
\author{
Sangeeta Bhatia
}
