% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_inc_with_corrections_DS1.R
\name{compute_inc_with_corrections_DS1}
\alias{compute_inc_with_corrections_DS1}
\title{Compute incidence from data stream 1 (after duplicate entries have been merged). 
This function corrects for decreasing cumulative case numbers and missing data.}
\usage{
compute_inc_with_corrections_DS1(merged_dat)
}
\arguments{
\item{merged_dat}{dataframe containing data with no duplicates i.e. one entry per date}
}
\value{
a dataframe containing incidence and cumulative incidence obtained from merged_dat.
}
\description{
Compute incidence from data stream 1 (after duplicate entries have been merged). 
This function corrects for decreasing cumulative case numbers and missing data.
}
\details{
### how do we correct cum incidence which goes down ###
1) identify pairs t, t+1 so that cumI(t) > cumI(t+1)
2) try to remove cumI(t), or cumI(t+1). 
    If one of those two options, and ONLY one "works", 
    i.e. makes cumI increasing for all steps from t-1 to t+2
    then remove (i.e. set to NA and change missing_data to 1) 
    the datapoint which makes it work when removed
3)  if the two options in 2) work, then remove (i.e. set to NA and change missing_data to 1) 
    both data points (t and t+1), the incidence should be increasing from t-1 to t+2
4) if none of the two options in 2) work, then remove (i.e. set to NA and change missing_data to 1) 
    both data points (t and t+1). Check if resulting incidence is increasing from t-1 to t+2. 
    If it is increasing from t-1 to t+2, that's it. 
    If it is not increasing from t-1 to t+2, apply the above rules (1-4)
    as if t'=t-1 and t'+1=t+2.
Question: how to deal with issues towards the end of the dataset (i.e. towards present days in real-time)
If t+1 is the last datapoint with data, do the following:
2') try to remove cumI(t), 
    if this "works", 
    i.e. makes cumI increasing for from t-1 to t+1
    then remove t (i.e. set to NA and change missing_data to 1) 
3') if the option in 2) did not work, then remove t (i.e. set to NA and change missing_data to 1) anyway
    and apply the above rules (2'-3')
    as if t'=t-1 and t'+1=t+1.
Question: how to deal with issues towards the start of the dataset 
In order to do this we add a 0 cumI at the start, which will NOT be used apart from for this correction
}
\examples{

#

}
\author{
Pierre Nouvellet and Anne Cori
}

