% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{deviance}
\alias{deviance}
\title{Deviance}
\usage{
deviance(I0, pij, R, si, x)
}
\arguments{
\item{I0}{T X N matrix of incidence}

\item{pij}{N X N.  Element i, j is the
probability of moving from i to j.}

\item{R}{either 1 X N matrix of reproduction numbers in each
location at t or T X N matrix of reproduction numbers in each
location at each time step from 1 to t.}

\item{si}{A vector drawn from serial interval
distribution. If the length of the vector is less than
T, it is padded with 0s.}

\item{x}{N X 1 vector of observed incidence at step t.}
}
\value{
deviance
}
\description{
Deviance at a given set of parameters.
Deviance is a measure of the model discrepancy and is defined as -2
times the log likelihood.
}
\details{
\deqn{D(y, \theta) = -2 logp(y|\theta)}
}
\references{
Gelman, Andrew, et al. Bayesian data analysis.
Chapman and Hall/CRC, 1995.
}
\author{
Sangeeta Bhatia
}
