% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{dic}
\alias{dic}
\title{Deviance Information Criterion}
\usage{
dic(I0, pij, R, si, x)
}
\arguments{
\item{I0}{list of matrices. Each element in the list is the incidence
matrix for a single simulation.}

\item{pij}{list of matrices. Each matrix is the probability of
movement for a single simulation. If the list is shorter than other
lists, it will be recycled.}

\item{R}{list of reproduction vectors. Same as for the other
parameters.}

\item{si}{list of serial intervals.}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
\deqn{DIC = 2\hat{D}_{avg}(y) - D_{\hat{\theta}}(y)}
}
\references{
Gelman, Andrew, et al. Bayesian data analysis. Chapman
and Hall/CRC, 1995.
}
\author{
Sangeeta Bhatia
}
