% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{disaggregate}
\alias{disaggregate}
\title{Return disaggregated counts}
\usage{
disaggregate(total, pmatrix, sim = 1000)
}
\arguments{
\item{total}{case count for large spatial unit}

\item{pmatrix}{matrix of probabilities. Each row corresponsd to a
single run and each column is a spatial sub-unit.}

\item{sim}{Number of simulations (i.e., draws) for the multinomial
distribution. Note that this need not be the same as the number of
rows in pmatrix.}
}
\value{
matrix with nrow(pmatrix) * sim rows and ncol(pmatrix)
columns
}
\description{
Distribute the case count for a spatial unit into subunits
}
\author{
Sangeeta Bhatia
}
