% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{disaggregate_distr}
\alias{disaggregate_distr}
\title{Quantiles for disaggregated count distribution}
\usage{
disaggregate_distr(sample_mat, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{sample_mat}{matrix of samples. Each row is a simulation and
each column is a spatial sub-unit. Make sure that the matrix has
meaningful column names so that the output makes sense.}
}
\value{
desired quantiles for each spatial sub-unit.
}
\description{
Quantiles for disaggregated count distribution
}
