% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_cleanup.R
\name{filter.case.count}
\alias{filter.case.count}
\title{Filter the case count for parameters of interest.}
\usage{
filter.case.count(case.count, species, disease, location)
}
\description{
Filters the input data on species, disease and location
}
\details{
.. content for \details{} ..
}
\author{
Sangeeta Bhatia
}
