% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_cleanup.R
\name{filter_case_count}
\alias{filter_case_count}
\title{Filter the case count for parameters of interest.}
\usage{
filter_case_count(case_count, species, disease, location)
}
\description{
Filters the input data on species, disease and location
}
\author{
Sangeeta Bhatia
}
