% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_processing.R
\name{gravity_model_flow}
\alias{gravity_model_flow}
\title{Computes the flow from A to B under the gravity model}
\usage{
gravity_model_flow(N_from, N_to, distance, K, pow_N_from, pow_N_to, pow_dist)
}
\arguments{
\item{N_from}{population of the source}

\item{N_to}{population of the destination}

\item{pow_N_from}{power on the population of the source}

\item{pow_N_to}{power on the population of the destination}

\item{pow_dist}{power on the distance between the source and the
destination}

\item{dist}{distance between the two places}
}
\value{
estimated flow between source and destination
}
\description{
Given the populations of A and B and the distance between them,
return the estimated population flow between
them modeled as
\phi(A,B) = K N_A^{\alpha}N_B^{\beta}/r_{AB}^{\gamma}..
}
\author{
Pierre Nouvellet, Anne Cori Sangeeta Bhatia
}
