% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incid_DS1.R
\name{incidence.from.DS1}
\alias{incidence.from.DS1}
\title{Compute incidence from data stream 1}
\usage{
incidence.from.DS1(case.count, species, disease, case.type = c("SCC", "SC",
  "CC", "SCD", "SD", "CD", "ALL"), location, merge_rule = c("median"))
}
\arguments{
\item{case.count}{dataframe containing data stream 1 to be processed}

\item{species}{string giving species in which disease incidence will be
extracted}

\item{disease}{string specifying disease of interest}

\item{case.type}{string specifying column name from which incidence is
extracted; one of "SCC" (suspected and confirmed cases), "SC" (suspected
cases), "CC" (confirmed cases), "SCD" (suspected and confirmed deaths),
"SD" (suspected deaths), or "CD" (confirmed deaths).}

\item{location}{string specifying location for which incidence will be
computed}

\item{merge_rule}{string to specify how merging should be made when
duplicated entries exist. At the moment only "median" is supported.}
}
\value{
a dataframe containing incidence and cumulative incidence obtained
  from merged_dat.
}
\description{
This function serves as a wrapper around the cleaning and processing of
incidence data. The functions for these are available in ...
}
\details{
XXX
}
\examples{

# to be written

}
\author{
Pierre Nouvellet, Anne Cori and Sangeeta Bhatia
}
