% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_cleanup.R
\name{interval.width.for.p}
\alias{interval.width.for.p}
\title{Interval width for Chebysev Inequality with sample mean.}
\usage{
interval.width.for.p(n, p)
}
\description{
.. For sample size n, what should k be so that the p% of the data is outside
k standard deviations of the sample mean ..
}
\details{
if k=interval.width.for.p(n, p), then chebyshev.ineq.sample(n, k) should be p.
}
\author{
Sangeeta Bhatia
}
