% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_cleanup.R
\name{interval_width_for_p}
\alias{interval_width_for_p}
\title{Interval width for Chebysev Inequality with sample mean.}
\usage{
interval_width_for_p(n, p)
}
\description{
For sample size n, what should k be so that the p% of the data
is outside k standard deviations of the sample mean.
}
\details{
if k=interval_width_for_p(n, p), then
chebyshev_ineq_sample(n, k) should be p.
}
\author{
Sangeeta Bhatia
}
