% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_processing.R
\name{is_outlier}
\alias{is_outlier}
\title{Tests for outlier.}
\usage{
is_outlier(cases, use_last, k_sd)
}
\arguments{
\item{cases}{numeric vector}

\item{use_last}{number of points preceeding the last one to be
used for estimation. If this is more than
the number of available data points, all points are used.}
}
\value{
TRUE is the last point is outside the prediction_interval
}
\description{
Test if the last data point is within the prediction interval
returned by prediction_interval
}
\details{
k_sd controls the proportion of the data that you want to
grab. If this value is large, greater variance
will be tolerated.
}
\seealso{
chebyshev_ineq_sample for the proportion *outside* a
certain interval around the mean
}
\author{
Sangeeta Bhatia
}
