% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Log likelihood under given model.}
\usage{
log_likelihood_t(I0, pij, R, si, x)
}
\arguments{
\item{I0}{T X N matrix of incidence}

\item{pij}{N X N.  Element i, j is the
probability of moving from i to j.}

\item{R}{either 1 X N matrix of reproduction numbers in each
location at t or T X N matrix of reproduction numbers in each
location at each time step from 1 to t.}

\item{si}{A vector drawn from serial interval
distribution. If the length of the vector is less than
T, it is padded with 0s.}

\item{x}{N X 1 vector of observed incidence at step t.}
}
\value{
log likelihood of observing x at step t.
}
\description{
Log-likelihood of observing the given incidence under the model.
}
\details{
The model is
\deqn{I_{j, t} = \sum_{i = 1}^N{p_{ij} R_{i, t} \sum_{s = 1}^t{I_{i, t - s}si_s}}}
}
\author{
Sangeeta Bhatia
}
