% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_cleanup.R
\name{merge.duplicates}
\alias{merge.duplicates}
\title{Merge dulplicate records in the incidence count.}
\usage{
\method{merge}{duplicates}(case.count, cols.to.keep, merge_rule = c("median"))
}
\description{
Duplicate records in the case count are merged according to the following rules.
Details
}
\author{
Sangeeta Bhatia
}
