% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_cleanup.R
\name{merge_duplicates}
\alias{merge_duplicates}
\title{Merge dulplicate records in the incidence count.}
\usage{
merge_duplicates(case_count, cols_to_keep, merge_rule = c("median"))
}
\value{
duplicates free data frame
}
\description{
Duplicate records in the case count are merged according to the
following rules.
Details
}
\author{
Sangeeta Bhatia
}
