% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{params_eff}
\alias{params_eff}
\title{Effective number of unconstrained parameters in the model}
\usage{
params_eff()
}
\description{
Effective number of unconstrained parameters in the model
}
\details{
\deqn{p_D^1 = \hat{D}_{avg}(y) - D_{\hat{\theta}}(y)}
}
\references{
Gelman, Andrew, et al. Bayesian data analysis. Chapman
and Hall/CRC, 1995.
}
\author{
Sangeeta Bhatia
}
