% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDS1.R
\name{processDS1}
\alias{processDS1}
\title{Process data from data stream 1}
\usage{
processDS1(dat, spec, disease, case_type = c("SCC", "SC", "CC", "SCD", "SD",
  "CD"), location, merge_rule = c("median"))
}
\arguments{
\item{dat}{dataframe containing data stream 1 to be processed}

\item{spec}{string giving species in which disease incidence will be extracted}

\item{disease}{string specifying disease of interest}

\item{case_type}{string specifying column name from which incidence is extracted; one of "SCC" (suspected and confirmed cases), "SC" (suspected cases), "CC" (confirmed cases), "SCD" (suspected and confirmed deaths), "SD" (suspected deaths), or "CD" (confirmed deaths).}

\item{location}{string specifying location for which incidence will be computed}

\item{merge_rule}{string to specify how merging should be made when duplicated entries exist. At the moment only "median" is supported.}
}
\description{
This function BLABLA
}
\details{
XXX
}
\examples{

# to be written

}
\author{
Pierre Nouvellet and Anne Cori
}

