% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_data_processing.R
\name{remove_last_outliers}
\alias{remove_last_outliers}
\title{Outlier Removal}
\usage{
remove_last_outliers(cum_incidence, use_last = 20, k_sd = 6)
}
\details{
If the last datapoint is an outlier, remove it
Do this until there is no outlier at the end.
}
\seealso{

}
\author{
Sangeeta Bhatia
}
