% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_projections.R
\name{sample_projections}
\alias{sample_projections}
\title{Draw random sample from a list of data frames}
\usage{
sample_projections(projections_list, place, k)
}
\arguments{
\item{projections_list}{list of data-frames. Each data frame is a
single simulation. The rows are dates and columns are places.}

\item{place}{name of column to be selected from each simulation}

\item{k}{number of samples to draw. If k > length(projections_list), it is
set to length(projections_list)}
}
\description{
Draw random sample from a list of data frames
}
\author{
Sangeeta Bhatia
}
