% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_projections.R
\name{shuffle_cols}
\alias{shuffle_cols}
\title{Shuffle columns of data frame}
\usage{
shuffle_cols(df)
}
\arguments{
\item{df}{data frame}
}
\value{
shuffled data frame
}
\description{
Shuffle the columns of a data frame
}
\details{
Each row of the data frame is shuffled *separately*
}
\author{
Sangeeta Bhatia
}
