% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{week_to_maxdate}
\alias{week_to_maxdate}
\title{Week of year to date conversion}
\usage{
week_to_maxdate(year, week)
}
\arguments{
\item{week}{a number between 1 and 53}
}
\value{
date
}
\description{
Convert a week of year string to the last date of the week
}
\details{
For our purposes, Week 1 of the year starts on 1st Jan
and ends on 7th Jan. I am aware that there is more to the logic than
this but for now, this simple rule will suffice. For week i, function
returns 7 * (i - 1) days have passed from the start of the year. Use
lubridate to add 7 * i to 1-1-year
}
\author{
Sangeeta Bhatia
}
