% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project future incidence at a location
Simulates future incidence based on past incidence data at a
given set of locations, the Reproduction numbers at each location and the distribution of the serial interval.
Runs a single simulation.}
\usage{
project(incid, R, si, pij, n.days = 7)
}
\arguments{
\item{incid}{A data frame of incidence counts for n locations and t days.
Dates run down the rows and locations are across columns. Must contain a column called Date.}

\item{R}{A positive vector of reproduction numbers for each location.}

\item{si}{A vector}

\item{pij}{a n x n matrix of probabilities where n is the number of locations.}

\item{n.days}{number of days for which each simulation should be run.}
}
\value{
data frame containing projected incidence count with n.days rows and n columns.
}
\description{
Project future incidence at a location
Simulates future incidence based on past incidence data at a
given set of locations, the Reproduction numbers at each location and the distribution of the serial interval.
Runs a single simulation.
}
\author{
Sangeeta Bhatia
}
